info.txt 4/2 1335
