+++ src/index.html
@@ -42,10 +42,13 @@
         class="gif"
       />
     </div>
-    Powered by:
-    <a href="https://www.weatherapi.com/" title="Free Weather API"
-      >WeatherAPI.com</a
-    >
+
+    <p class="weather-api-link">
+      Powered by:
+      <a href="https://www.weatherapi.com/" title="Free Weather API"
+        >WeatherAPI.com</a
+      >
+    </p>
     <div class="footer"></div>
   </body>
 </html>
