+++ src/ui.js
@@ -3,65 +3,39 @@
 //ui.js
 
 //Imports
-import { fetchWeather } from "./weather.js";
-import {
-    locationForm,
-    currentLocation,
-    currentCondition,
-    currentTemp,
-    radioButtonF,
-    radioButtonC,
-    gif,
-    defaultGifSearch,
-    weatherApiKey,
-    defaultLocation,
-    gifyApiKey,
-    currentLocationUrl,
-} from './website.js';
+import { fetchWeather,
+} from "./weather.js";
 
 // //Declarations
-// const locationForm = document.getElementById("search-location");
-// const currentLocation = document.querySelector(".current-location");
-// const currentCondition = document.querySelector(".current-condition");
-// const currentTemp = document.querySelector(".current-temp");
-// const radioButtonF = document.getElementById("fahrenheit"); //can this be consolidated into queryselectorall and still function
-// const radioButtonC = document.getElementById("celsius");
-// const gif = document.querySelector(".gif");
-// const defaultGifSearch = "random weather";
-// const weatherApiKey = "9a15cd9d2f10400dbc7152440242003";
-// const defaultLocation = `https://api.weatherapi.com/v1/current.json?key=${weatherApiKey}&q=84129`;
-// const gifyApiKey = "OLRY4DR4fNBTQXbcI1NBFibXOW6q39k7";
+export const locationForm = document.getElementById("search-location");
+export const currentLocation = document.querySelector(".current-location");
+export const radioButtonF = document.getElementById("fahrenheit");
+export const radioButtonC = document.getElementById("celsius");
+export const weatherApiKey = "9a15cd9d2f10400dbc7152440242003";
+export const defaultLocation = `https://api.weatherapi.com/v1/current.json?key=${weatherApiKey}&q=84129`;
+export let currentLocationUrl = defaultLocation;
 
-//eventListeners
-
+//Event Listeners
 export function attachEventListeners() {
-// export function attachEventListeners() {
   radioButtonF.addEventListener("change", function () {
     if (this.checked) {
-      // const location = displayLocation;
-    //   fetchWeather(displayLocation, "fahrenheit");
-    fetchWeather(currentLocationUrl, "fahrenheit")
+      fetchWeather(currentLocationUrl, "fahrenheit");
     }
   });
 
   radioButtonC.addEventListener("change", function () {
     if (this.checked) {
-      // const location = displayLocation;
-    //   fetchWeather(displayLocation, "celsius");
-    fetchWeather(currentLocationUrl, "celsius");
+      fetchWeather(currentLocationUrl, "celsius");
     }
   });
 
   locationForm.addEventListener("submit", function (e) {
     e.preventDefault();
-    const location = document.getElementById("location-input").value;
-    // const weatherApiUrl = `https://api.weatherapi.com/v1/current.json?key=${weatherApiKey}&q=${encodeURIComponent(
-    //   location
-    // )}`;
-    const newLocationUrl = `https://api.weatherapi.com/v1/current.json?key=${weatherApiKey}&q=${encodeURIComponent(location)}`;
-    currentLocationUrl = newLocationUrl;
-    // fetchWeather(weatherApiUrl, "fahrenheit");
-    fetchWeather(newLocationUrl, "fahrenheit");
+    const location = document.getElementById("location-input").value; 
+    currentLocationUrl = `https://api.weatherapi.com/v1/current.json?key=${weatherApiKey}&q=${encodeURIComponent(
+      location
+    )}`;
+    fetchWeather(currentLocationUrl, "fahrenheit");
     document.getElementById("location-input").value = "";
   });
 }
